[minor]: makes the code run properly =)
This goes along with the changes done in `cell.py` and `network.py` in
c7b9c201e560f57d5954338c0f714868d06360f6.

[Still... the code doesn't seem to be learning greatly. We should (1)
try tying the weights; and (2) check what happens when we change the
loss function. Also, there are a lot of hyperparamenters, and we are
training with `batch_size = 1`, which is not great]

 Changes to be committed:
	modified:   train.py