"Supposedly" implement tied weights (which don't work)
There is some conceptual problem I realized only after starting this
implementation: the size of the first layer is different from that of
the other layers, so it doesn't make sense to talk about "tying"
weights (since, after all, they simply can't be the same). But I may
be missing something, so I need to discuss this with someone else.

How is this implemented? The Network class keeps a reference of the
previous cell and pass this reference to the Cell class constructor.
If the constructor receives this reference, it initializes its `wf`,
`bf`, `wi`, `bi`, (and so on) with the variables present in the
previous cell. *I AM ASSUMING* that doing this will cause TensorFlow
to treat the two variables as if they were the same. I didn't try this
in a toy example.

Besides this implementation, there are some minor changes:

* Putting spaces after ','
* Putting spaces around '='
* Making the code Python3 compliant (but also still Python2) (I needed
	this because my TensorFlow is installed for Python3).

 Changes to be committed:
	modified:   cell.py
	modified:   network.py
	modified:   train.py